Code refactor based on clang tidy
- added GSL library (guidelines support library)
- removed some of the checks that were adding too much noise
- partial code refactor

+++ src/Missle.cpp
@@ -2,10 +2,11 @@
 
 #include <SFML/Graphics.hpp>
 #include <cmath>
+#include <gsl/gsl>
 
 #include "Size.hpp"
 
-Missle::Missle(sf::Texture& texture, const int angle, const float x, const float y)
+Missle::Missle(sf::Texture& texture, const float angle, const float x, const float y)
     : mPos({x, y}), mAngle(angle) {
   mSprite.setTexture(texture);
   mSprite.setPosition(mPos.x, mPos.y);
@@ -23,6 +24,6 @@ void Missle::update() {
   const auto rotation_degree = mAngle - 90;
   const auto rotation_radians = PI / 180.f * rotation_degree;
 
-  mPos.x += mSpeed * std::cos(rotation_radians);
-  mPos.y += mSpeed * std::sin(rotation_radians);
+  mPos.x += mSpeed * gsl::narrow_cast<float>(std::cos(rotation_radians));
+  mPos.y += mSpeed * gsl::narrow_cast<float>(std::sin(rotation_radians));
 }
